requirements.txt
Django==2.0.3
gunicorn==19.8.1
psycopg2==2.7.4
django-heroku==0.3.1
djangorestframework==3.8.2
